Allow disabling AGENT

+++ docker/startup.sh
@@ -1,16 +1,19 @@
 #!/bin/sh
 
-echo "Downloading agent from ${AGENT_URI}"
-wget -T 10 -O agent.zip "${AGENT_URI}"
+# shellcheck disable=SC2039
+if [ "$WITH_AGENT" = true ] ; then
+  echo "Downloading agent from ${AGENT_URI}"
+  wget -T 10 -O agent.zip "${AGENT_URI}"
 
-if [ -f agent.zip ] ; then
-  echo "Cleaning up agent..."
-  rm -fr /opt/agent-distribution
+  if [ -f agent.zip ] ; then
+    echo "Cleaning up agent..."
+    rm -fr /opt/agent-distribution
 
-  echo "Unzip agent compress file..."
-  unzip agent.zip
-else
-  echo "Failed to downloading agent..."
+    echo "Unzip agent compress file..."
+    unzip agent.zip
+  else
+    echo "Failed to downloading agent..."
+  fi
 fi
 
 if [ -f /opt/shared/conf/jvm.config ] ; then
